Allow lazy loading of locators in web actions
Extend the handling of web Actions so that the Locator can be provided
via a function. This allows for the lazy loading of Loacator objects
based on the Playwright Page provided at execution time. This allows for
the use of advanced features in Playwright for selecting elements, such
as the getByRole() function, which could not as easily used previously.

Refs:
- https://github.com/testla-project/testla-screenplay-playwright-js/issues/66